WEEK 4
This week the main components that were added in included a resume (which has my personal details redacted), and a form where people can contact me with queries. I also did some cleaning of the website code and making sure to update my comments on the website, and editing my formatting to make the overall visual appeal of the website more cohesive. This included editing the CSS file to make the words in an unordered list and the words in a table a black colour as they were previously grey.  I also made sure to add sections to each page so the text had a white background and appeared in a box.