Add files via upload
This week I did a lot of editing my HTML code and making sure that there were comments where needed and fixing spelling mistakes! I also added a place for my form responses to be collected, and tested this and saw my response on my email that I have linked to this form.